Drop packets while the WAN is up
When the device on the LAN side doesn't get an IP address, it attempts
to resetart the authentication. This causes the WAN side to go down
while reauthentication happens. We want to stop this.

Currently, we attempt to see if the WAN interface has an IP address.
While it does, we drop packets. This means that later on if the WAN
does go down, we can start forwarding packets again and bring the WAN
back up.

A better, but slightly more difficult approach, would be to inspect
the 802.1x packets themselves to determine when reauthentication is
being requested and drop only these packets. Patches for this would be
welcome.